# apps/api/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Corporate Action Guard API",
    "version": "1.0.0",
    "description": "Fail-closed preflight authorization for integrations acting on xStocks corporate-action state. A receipt is issued only for an ALLOW decision. The guard is enforceable only for paths that route through ActionGuardAdapter; a direct ERC-20 transfer bypasses it. X Layer mainnet is read-only.",
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Format cag_<keyId>_<secret>. Only the sha256 hash is stored server-side; the raw key is never logged."
      }
    },
    "schemas": {
      "Problem": {
        "type": "object",
        "required": ["type", "title", "status"],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "correlationId": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
      "ReviewResolution": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "minLength": 20,
            "maxLength": 2000
          },
          "evidenceIds": {
            "minItems": 1,
            "maxItems": 100,
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          }
        },
        "required": ["reason", "evidenceIds"],
        "additionalProperties": false
      }
    }
  },
  "paths": {
...
    "/v1/assets": {
      "get": {
        "tags": ["evidence"],
        "summary": "List discovered assets with lifecycle state, canonicality, and evidence freshness.",
        "description": "Keyset pagination. Offset pagination would silently skip or repeat rows as projections update underneath a paging client, which for an evidence console means an asset in a bad state could scroll past unseen. Every response carries servedAt.",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 25
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Opaque; from a prior nextCursor."
          },
          {
            "name": "lifecycleState",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "canonicality",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["PASS", "FAIL", "UNKNOWN"]
            }
          },
          {
            "name": "staleEvidence",
            "in": "query",
            "schema": {
              "type": "boolean"
            },
            "description": "True selects stale or missing evidence; false selects fresh evidence."
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 64
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of assets."
          },
          "400": {
            "description": "An unrecognised filter is rejected rather than ignored.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/assets/{assetId}": {
      "get": {
        "tags": ["evidence"],
        "summary": "One asset with its current canonicality and evidence provenance.",
        "parameters": [
          {
            "name": "assetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The asset."
          },
          "404": {
            "description": "No such asset. An empty shell is never returned in its place.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/assets/{assetId}/timeline": {
      "get": {
        "tags": ["evidence"],
        "summary": "Replay one asset from immutable journal events.",
        "parameters": [
          {
            "name": "assetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "upToEventId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500,
              "default": 200
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Journal events and deterministic replay metadata."
          },
          "400": {
            "description": "Invalid or unrecognised timeline query parameters.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "No such asset.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/incidents": {
      "get": {
        "tags": ["evidence"],
        "summary": "Open and resolved incidents, ordered by deterministic severity then recency.",
        "description": "Severity is a deterministic category. There is no numeric or model-generated risk score.",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["OPEN", "IN_REVIEW", "RESOLVED", "RECOVERED"]
            }
          },
          {
            "name": "assetId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 25
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Incidents."
          }
        }
      }
    },
    "/v1/preflight": {
      "post": {
        "tags": ["preflight"],
        "summary": "Authorize one exact operation, or refuse it.",
        "description": "Returns ALLOW with a short-lived EIP-712 receipt bound to this exact operation, or BLOCK with stable reason codes and NO receipt. A receipt is never returned for a BLOCK, for degraded evidence, or for an UNKNOWN check. An Idempotency-Key header is mandatory: a retried preflight must not mint a second receipt for the same intent.",
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 128
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreflightRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "An ALLOW with a receipt, or a BLOCK with reason codes.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreflightResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request, or a missing Idempotency-Key.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed. The body is identical for every failure reason.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "The key lacks the integrator:preflight scope.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "The idempotency key is in progress or was reused for a different request.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/incidents/{incidentId}/review-resolution": {
      "post": {
        "tags": ["operator"],
        "summary": "Record an operator resolution. Does not resume protected actions.",
        "description": "A resolution records a decision with an actor, a specific reason, and an evidence reference. It cannot create source agreement: the response always reports protectedActionsResumed=false, and protected actions remain governed by current evidence.",
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "name": "incidentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewResolution"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Recorded. Protected actions are not resumed."
          },
          "400": {
            "description": "A specific reason and an evidence reference are mandatory.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "The key lacks the operator:review scope.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "No such incident.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "The incident was already resolved.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    }
  }
}